Create Pi Estimation
One of the many things I remember in my teaching experience is explaining the concept of pi to students prior to the introduction of topics like mensuration. This code takes you through the approximation of pi.
This involves generating a large number of random points, and then counting how many of those points fall within a quarter-circle (or unit circle) inscribed within the square. The ratio of the number of points inside the quarter-circle to the total number of points will be an approximation of pi/4.